+++ src/commands/misc/qrcode.js
@@ -8,8 +8,10 @@ module.exports = {
     description: "Generate/Read a qrcode",
     usage: "[subcommand] [data / image_url]",
     subcommand: true,
-    subcommands: ["generate", "read"],
-    subs_desc: ["Generate a qrcode", "Read a qrcode"],
+    subcommands: [
+        { name: "generate", desc: "Generate a qrcode" },
+        { name: "read", desc: "Read a qrcode" },
+    ],
     cooldown: 10,
     category: "Miscellaneous",
     async execute(message, args, guildDB) {