+++ src/commands/mod/prune.js
@@ -7,19 +7,25 @@ const { Permissions } = require("discord.js");
 module.exports = {
     name: "prune",
     description: "Prune messages.",
+    aliases: ["purge"],
     permissions: [Permissions.FLAGS.MANAGE_GUILD],
     bot_perms: [Permissions.FLAGS.MANAGE_MESSAGES],
     args: true,
     guildOnly: true,
     usage: "[no of msg to prune / subcommand]",
-    subcommands: ["all", "bots", "*[string]"],
-    subs_desc: [
-        "Delete 100 messages",
-        "Delete all messages sent by a bot",
-        '`*Text` will delete any message containing "Text"',
+    subcommands: [
+        { name: "all", desc: "Delete 100 messages" },
+        {
+            name: "bots",
+            desc: "Delete all messages sent by a bot in this channel",
+        },
+        {
+            name: "*[string]",
+            desc: '`*Text` will delete any message containing "Text"',
+        },
     ],
-    cooldown: 10,
-    category: "Manage",
+    cooldown: 5,
+    category: "Moderation",
     execute(message, args, guildDB) {
         let messages;
         let errMsg =
@@ -40,8 +46,9 @@ module.exports = {
                 (m) => m.content.indexOf(args[0]) !== -1
             );
         }
+        message.delete();
         if (!isNaN(parseInt(args[0]))) {
-            const amount = parseInt(args[0]) + 1;
+            const amount = parseInt(args[0]);
             if (isNaN(amount)) {
                 return message.reply(
                     "The provided number of messages to delete doesn't seem to be a valid number."
@@ -67,7 +74,7 @@ module.exports = {
             }
         }
         message.channel
-            .send("Pruning done👍. This message will be deleted in 5 seconds")
+            .send("Pruning done 👍. This message will be deleted in 5 seconds")
             .then((msg) => {
                 setTimeout(() => {
                     msg.delete();