+++ src/functions/execute.js
@@ -76,7 +76,7 @@ module.exports = async (message, guildDB) => {
             const botPerms = message.guild.me.permissionsIn(message.channel);
             if (!botPerms || !botPerms.has(command.bot_perms)) {
                 return message.reply(
-                    `You didn't give the bot permission to do this!\nSend \`${guildDB.prefix}help ${command.name}\` to get list of permissions required by this command.`
+                    `You didn't give the bot permission to do this!\nSend \`${guildDB.prefix}help ${command.name}\` to get list of permissions required by this command.\nDon't know what you have given already? Send \`${guildDB.prefix}botperms\` in this channel itself.`
                 );
             }
         }
@@ -104,7 +104,11 @@ module.exports = async (message, guildDB) => {
             let reply = `Subcommands are required for this command.`;
 
             if (command.subcommands) {
-                reply += `\nThe subcommand(s) available are: \`${command.subcommands.join(
+                let subcmds = [];
+                for (var i = 0; i < command.subcommands.length; i++) {
+                    subcmds.push(command.subcommands[i].name);
+                }
+                reply += `\nThe subcommand(s) available are: \`${subcmds.join(
                     ", "
                 )}\``;
             }
@@ -170,7 +174,6 @@ module.exports = async (message, guildDB) => {
                 message.reply({ embeds: [embed] });
                 return;
             }
-            return true;
         }
     } else if (message.content.startsWith(guildDB.prefix.trim())) {
         //message.reply(errMsg);