+++ src/helpers/updateDocs.js
@@ -8,9 +8,6 @@ module.exports = (client) => {
     const fs = require("fs");
     const { commands, categories } = client;
     let text = fs.readFileSync(__dirname + "/cmdTemplate.md", "utf8");
-    text = text
-        .replace("{commandsRoundToTen}", `${Math.floor(commands.size / 10)}0`)
-        .replace("{categoriesSize}", categories.length);
     let toc = "# Table of contents\n\n"; //Table of contents
     categories.forEach((cat) => {
         const cmds = commands
@@ -19,30 +16,34 @@ module.exports = (client) => {
         toc += `- [${cat.name}](#${cat.name
             .toLowerCase()
             .replace(" ", "-")})\n`;
-        text += `## ${cat.name} (${cmds.length} commands)\n\n`;
+        text += `\n## ${cat.name} (${cmds.length} commands)\n`;
         cmds.forEach((cmd) => {
             let subcommands;
             if (cmd.subcommands) {
                 subcommands = [];
                 for (var i = 0; i < cmd.subcommands.length; i++) {
                     subcommands.push(
-                        `\`${cmd.subcommands[i]}\` - ${cmd.subs_desc[i]}`
+                        `\`${cmd.subcommands[i].name}\` - ${cmd.subcommands[i].desc}`
                     );
                 }
             }
             text +=
-                `### \`${cmd.name}\`\n\n` +
+                `\n### \`${cmd.name}\`\n\n` +
                 `##### Subcommands:\n\n- ${
                     subcommands ? subcommands.join("\n- ") : "None"
-                }\n\n##### Info\n\n` +
+                }\n\n##### Cmd info\n\n` +
                 `- Description: ${cmd.description}\n` +
                 `- Usage: ${cmd.usage ? cmd.usage : "None"}\n` +
                 `- Aliases: ${
-                    cmd.aliases ? cmd.aliases.join(", ") : "None"
+                    cmd.aliases ? `\`${cmd.aliases.join("`, `")}\`` : "None"
                 }\n` +
-                `- Cooldown: ${cmd.cooldown}\n\n`;
+                `- Cooldown: ${cmd.cooldown}\n`;
         });
     });
-    fs.writeFileSync("./commands.md", `${toc}\n${text}`);
+    text = text
+        .replace("{commandsRoundToTen}", `${Math.floor(commands.size / 10)}0`)
+        .replace("{categoriesSize}", categories.length)
+        .replace("{toc}", toc);
+    fs.writeFileSync("./commands.md", `${text}`);
     client.logger.log("Docs updated!", "debug");
 };